requirements.txt
boto==2.49.0
Django==3.0.3
djangorestframework==3.11.0
mysqlclient==1.3.14
djangorestframework-simplejwt==4.4.0